+++ settings/init_settings.py
@@ -4,6 +4,7 @@ from tempfile import gettempdir
 from re import sub, compile,I
 from os import path, mkdir
 from json import dumps,dump,load
+from sys import argv
 
 try: 
 	vbox_template = {
@@ -25,32 +26,52 @@ try:
 			]
 		  }
 
-	reg = compile(r"[^A-Za-z0-9\!\@\#\-\_ ]",I)
-	temp_location = path.join(gettempdir(),"rhinotempfiles")
-	if not path.exists(temp_location):
-		mkdir(path.join(gettempdir(),"rhinotempfiles"))
+	reg = compile(r"[^A-Za-z0-9\!\@\#\-\_\/]",I)
+	temp_location = sub(reg, "", path.join(gettempdir(),"rhinotempfiles"))
 	print("\nAdding virtual boxes settings (Allowed characters are: A-Z a-z 0-9_-!@#)\n")
-	n_boxes = raw_input("How many VMs do you have? ")
-	temp_location = raw_input("Default temporary folder is {} -> ".format(temp_location)) or temp_location
-	_dict_of_boxes = {}
-	for x in range(0,int(n_boxes)):
-		name = "vbox_{}".format(x)
-		print("\n----------{}----------\n".format(name))
-		temp_vbox_template = vbox_template
-		temp_vbox_template["vm"] = sub(reg, "", raw_input("VM name? (E.g. Ubuntu18, TestUbuntu or Windows10) "))
-		temp_vbox_template["snapshot"] = sub(reg, "", raw_input("VM snapshot name? (E.g. Snapshot_1) "))
-		temp_vbox_template["user"] = sub(reg, "", raw_input("VM OS username? (E.g. IEUser, user, Test) "))
-		temp_vbox_template["pass"] = sub(reg, "", raw_input("VM OS password? (E.g. easyPassw0rd!@) "))
-		temp_vbox_template["os"] = sub(reg, "", raw_input("VM OS Windows or Linux? (case-sensitive) "))
-		temp_vbox_template["queue"] = name
-		temp_vbox_template["temp"] = temp_location
+	if argv[1] != "dummy":
+		n_boxes = raw_input("How many VMs do you have? ")
+		temp_location = sub(reg, "", raw_input("Default temporary folder is {} -> ".format(temp_location))) or temp_location
+		if not path.exists(temp_location):
+			mkdir(temp_location)
+		_dict_of_boxes = {}
+		for x in range(0,int(n_boxes)):
+			name = "vbox_{}".format(x)
+			print("\n----------{}----------\n".format(name))
+			temp_vbox_template = vbox_template
+			temp_vbox_template["vm"] = sub(reg, "", raw_input("VM name? (E.g. Ubuntu18, TestUbuntu or Windows10) "))
+			temp_vbox_template["snapshot"] = sub(reg, "", raw_input("VM snapshot name? (E.g. Snapshot_1) "))
+			temp_vbox_template["user"] = sub(reg, "", raw_input("VM OS username? (E.g. IEUser, user, Test) "))
+			temp_vbox_template["pass"] = sub(reg, "", raw_input("VM OS password? (E.g. easyPassw0rd!@) "))
+			temp_vbox_template["os"] = sub(reg, "", raw_input("VM OS Windows or Linux? (case-sensitive) "))
+			temp_vbox_template["queue"] = name
+			temp_vbox_template["temp"] = temp_location
 
+			for key in temp_vbox_template:
+				if temp_vbox_template[key] == "":
+					print("{} is Empty {}".format(key,temp_vbox_template[key]))
+					exit()
+			_dict_of_boxes[name] = temp_vbox_template
+	else:
+		name = "vbox_1"
+		_dict_of_boxes = {}
+		if not path.exists(temp_location):
+			mkdir(temp_location)
+		_dict_of_boxes = {}
+		temp_vbox_template = vbox_template
+		temp_vbox_template["vm"] = "dummy"
+		temp_vbox_template["snapshot"] = "dummy"
+		temp_vbox_template["user"] = "dummy"
+		temp_vbox_template["pass"] = "dummy"
+		temp_vbox_template["os"] = "dummy"
+		temp_vbox_template["queue"] = "vbox_testing"
+		temp_vbox_template["temp"] = temp_location
 		for key in temp_vbox_template:
 			if temp_vbox_template[key] == "":
 				print("{} is Empty {}".format(key,temp_vbox_template[key]))
 				exit()
 		_dict_of_boxes[name] = temp_vbox_template
-
+	
 	print("\nAdding your settings to settings.json, you can manually edit all the settings later on\n")
 	print(dumps(_dict_of_boxes, indent=4, sort_keys=True))