+++ changes.md
@@ -1,3 +1,9 @@
+#### 2020.V.01.08
+- Fixed some backend issues
+- Added testing logic (create, restore and start)
+- Added one click dummy initialization
+- Added some fixes to auto switch
+
 #### 2020.V.01.07 (It took forever to come up with work around for these)
 - Most logics run from containers
 - Added many auto fixes and auto switching for handling VMs